requirements.txt
discord.py == 2.3.2
beautifulsoup4 == 4.12.2
pytz == 2023.3.post1
lxml == 4.9.3
requests == 2.31.0